Runs classification in a background job
Note:
classification_id was remove from the response,
because the classification method no longer return a result
but an Async Object for tracking job status
