+++ ml_classifier.py
@@ -3,7 +3,6 @@
 
 from flask import Flask, request, jsonify, make_response, abort
 from sklearn.feature_extraction.text import TfidfVectorizer
-from sklearn.externals import joblib
 from bs4 import BeautifulSoup
 import re
 import os
@@ -14,6 +13,7 @@ sys.path.append(os.path.dirname(__file__))
 from profanity_filter import *
 from nltk.corpus import stopwords
 import elasticsearch
+from tasks import evaluate_petition
 
 
 app = Flask(__name__)
@@ -31,25 +31,17 @@ def review_words(raw_text):
     return u" ".join(meaningful_words)
 
 
-def evaluate_petition(features):
-    pipe = joblib.load('models/bow_ng6_nf9500_ne750.pkl')
-    test_data_features = pipe.named_steps['vectorizer'].transform(features)
-    test_data_features = test_data_features.toarray()
-    result = pipe.named_steps['forest'].predict(test_data_features)
-    return result
-
-
 def get_relevant_hits(like_text):
     index_name = "peticion"
     doc_type = "pregunta"
     stop_words = ["para","apoyo","una","la","el","de","en"]
     body = {"query": { "more_like_this" : { "fields":["titulo"],
-            "like_text" : like_text, "min_term_freq" : 1, 
+            "like_text" : like_text, "min_term_freq" : 1,
             "max_query_terms" : 100, "min_doc_freq": 0,
             "stop_words":stop_words}}}
     es = elasticsearch.Elasticsearch()
     mlts = es.search(index=index_name, doc_type=doc_type,body=body)
-    relevant_sugestions = []                    
+    relevant_sugestions = []
     hits = mlts.get('hits')["hits"]
     for hit in hits:
         if hit["_score"] >= 0.5:
@@ -57,7 +49,7 @@ def get_relevant_hits(like_text):
             relevant_sugestions.append({"title": source["titulo"],
             "description": source["sugerencia"], "link": source["link"]})
     return relevant_sugestions
-    
+
 
 @app.errorhandler(400)
 def bad_request(error):
@@ -98,12 +90,11 @@ def create_task():
     clean_test_descripciones.append(u" ".join(
         KaggleWord2VecUtility.review_to_wordlist(features, True))
         )
-    myeval = evaluate_petition(clean_test_descripciones)
+    myeval = evaluate_petition.delay(clean_test_descripciones)
     # tasks.append(task)
     return jsonify({'id': request.json['id'],
                     'title': request.json['title'],
-                    'description': request.json['description'],
-                    'classification_id': myeval[0]}), 201
+                    'description': request.json['description']}), 201
 
 
 @app.route('/recommendations', methods=['GET'])
@@ -115,7 +106,7 @@ def get_hits():
     }
     relevant_sugestions = get_relevant_hits(task['title'])
     return jsonify({'results': relevant_sugestions})
-    
-    
+
+
 if __name__ == '__main__':
     app.run(host='0.0.0.0', debug=True)
